refactor(#998): relocate isNonCompliant to compliance.ts (architect nit)
Architect review: isNonCompliant is a compliance-policy predicate (takes
a vehicle shape, encodes 'operator must act') and belongs beside its
siblings isRoadLegal/isDocCurrent in compliance.ts, not in the date-
primitive module expiry.ts. compliance.ts already imports
computeExpiryStatus, so the move adds no new coupling and keeps the clean
one-way edge; expiry.ts goes back to a pure date primitive. Tests follow
the predicate into compliance.test.ts. fleet-filters imports from
@kuruma/shared/lib/compliance. Also reworded the now-relocated banner's
'On the operator dashboard' comment (code-review LOW).

Refs #998
