+++ packages/shared/src/lib/expiry.ts
@@ -14,28 +14,3 @@ export function computeExpiryStatus(expiryDate: string | null, todayIso: string)
   if (expiryDate <= thresholdIso) return 'EXPIRING_SOON'
   return 'OK'
 }
-
-/** One document is expiring-soon or already expired as of `todayIso`. A missing
- *  (UNKNOWN) date is excluded — see `isNonCompliant`. */
-function needsAttention(expiryDate: string | null, todayIso: string): boolean {
-  const status = computeExpiryStatus(expiryDate, todayIso)
-  return status === 'EXPIRING_SOON' || status === 'EXPIRED'
-}
-
-/**
- * Whether a vehicle needs the operator's compliance attention: its shaken OR
- * insurance certificate is expiring within `EXPIRY_SOON_DAYS` or already expired.
- * The named rule behind the dashboard banner, the fleet `expiringSoon` facet, and
- * the fleet filter, so all three count the same set. A MISSING (null) date is NOT
- * counted — whether legacy null-doc cars should surface here is the open #998
- * item-2 product call; this is the single place to change it.
- */
-export function isNonCompliant(
-  vehicle: { shakenExpiryDate: string | null; insuranceExpiryDate: string | null },
-  todayIso: string,
-): boolean {
-  return (
-    needsAttention(vehicle.shakenExpiryDate, todayIso) ||
-    needsAttention(vehicle.insuranceExpiryDate, todayIso)
-  )
-}
